Adds Rectangle and Buffer wrapper classes
Both classes are missing a lot of methods,
but the underlying gir objects can be
acessed as attributes of these objects.
The main goal here was to have a way
to access GEGL Pixel data from within
Python itself: that can be done
using the gegl:write-buffer op,
and the "get" method on gegl.Buffer.